Fix: specs, add database clean politics between transactions and random order

+++ spec/app_id_service_spec.rb
@@ -1,7 +1,19 @@
 # frozen_string_literal: true
 
+require 'rails_helper'
+
 RSpec.describe AppIdService do
   describe '#version' do
+    before do
+      # Reset cached version before each test
+      described_class.instance_variable_set(:@version, nil)
+    end
+
+    after do
+      # Reset cached version after each test to avoid polluting other tests
+      described_class.instance_variable_set(:@version, nil)
+    end
+
     context 'when we use capistrano' do
       let(:hash) { SecureRandom.hex }
 
@@ -10,11 +22,21 @@ RSpec.describe AppIdService do
       end
 
       after do
-        File.delete('REVISION')
+        File.delete('REVISION') if File.exist?('REVISION')
       end
 
       it 'is equal the hash in REVISION file' do
         expect(described_class.version).to eq(hash.first(8))
+      end
+    end
+
+    context 'when REVISION file does not exist' do
+      before do
+        File.delete('REVISION') if File.exist?('REVISION')
+      end
+
+      it 'returns git short hash' do
+        expect(described_class.version).to match(/\A[a-f0-9]{7,8}\z/)
       end
     end
   end
